requirements.txt
Flask==2.3.2
Flask-PyMongo==2.3.0
Flask-Login==0.6.3
python-dotenv==1.0.0
pymongo==4.6.1
boto3==1.34.48
Werkzeug==3.0.1